requirements.txt
# Análisis Mercado Inmobiliario Sevilla — Grupo Insur
# Python 3.10+

# Datos y manipulación
pandas>=2.0
numpy>=1.24
xlrd>=2.0          # lectura de archivos .xls (MIVAU)
openpyxl>=3.1      # lectura de archivos .xlsx

# Extracción PDF
pdfplumber>=0.10

# API GenAI
anthropic>=0.25    # Claude API
python-dotenv>=1.0 # carga de ANTHROPIC_API_KEY desde .env

# Machine Learning
scikit-learn>=1.3

# Forecasting
statsmodels>=0.14  # SARIMA / SARIMAX
prophet>=1.1       # Prophet (requiere pystan o cmdstanpy)

# Visualización
plotly>=5.18
